+++ backend/src/certificates/certificates.module.ts
@@ -1,8 +1,10 @@
-import { Module } from '@nestjs/common'
+import { Module, forwardRef } from '@nestjs/common'
 import { CertificatesController } from './certificates.controller'
 import { CertificatesService } from './certificates.service'
+import { QueuesModule } from '../queues/queues.module'
 
 @Module({
+  imports: [forwardRef(() => QueuesModule)],
   controllers: [CertificatesController],
   providers: [CertificatesService],
   exports: [CertificatesService],